+++ R/df_printer.R
@@ -246,6 +246,13 @@ as_flextable.data.frame <- function(x,
                                     show_coltype = TRUE,
                                     color_coltype = "#999999",
                                     ...) {
+
+  if (inherits(x, "data.table")) {
+    x <- as.data.frame(x)
+  } else if (inherits(x, "tbl_df")) {
+    x <- as.data.frame(x)
+  }
+
   if (nrow(x) == 1) {
     singlerow_df_printer(
       dat = x,